v2.3.4: The mouse stuff and interface button positions should *actually* be fixed now.

+++ source/windowing/StarCanvasWidget.cpp
@@ -88,7 +88,7 @@ bool CanvasWidget::sendEvent(InputEvent const& event) {
     return false;
 
   auto& context = GuiContext::singleton();
-  int interfaceScale = m_ignoreInterfaceScale ? 1 : context.interfaceScale();
+  float interfaceScale = m_ignoreInterfaceScale ? 1.0f : context.interfaceScale();
   if (auto mouseButtonDown = event.ptr<MouseButtonDownEvent>()) {
     if (inMember(*context.mousePosition(event, interfaceScale)) && m_captureMouse) {
       m_clickEvents.append({*context.mousePosition(event, interfaceScale) - screenPosition(), mouseButtonDown->mouseButton, true});
